+++ apollo-client/src/main/java/com/ctrip/framework/apollo/spring/boot/ApolloApplicationContextInitializer.java
@@ -18,12 +18,12 @@ package com.ctrip.framework.apollo.spring.boot;
 
 import com.ctrip.framework.apollo.Config;
 import com.ctrip.framework.apollo.ConfigService;
+import com.ctrip.framework.apollo.core.ApolloClientSystemConsts;
 import com.ctrip.framework.apollo.core.ConfigConsts;
 import com.ctrip.framework.apollo.core.utils.DeferredLogger;
 import com.ctrip.framework.apollo.spring.config.ConfigPropertySourceFactory;
 import com.ctrip.framework.apollo.spring.config.PropertySourcesConstants;
 import com.ctrip.framework.apollo.spring.util.SpringInjector;
-import com.ctrip.framework.apollo.util.factory.PropertiesFactory;
 import com.google.common.base.Splitter;
 import com.google.common.base.Strings;
 import java.util.List;
@@ -77,9 +77,15 @@ public class ApolloApplicationContextInitializer implements
   public static final int DEFAULT_ORDER = 0;
 
   private static final Logger logger = LoggerFactory.getLogger(ApolloApplicationContextInitializer.class);
-  private static final Splitter NAMESPACE_SPLITTER = Splitter.on(",").omitEmptyStrings().trimResults();
-  private static final String[] APOLLO_SYSTEM_PROPERTIES = {"app.id", ConfigConsts.APOLLO_CLUSTER_KEY,
-      "apollo.cacheDir", "apollo.accesskey.secret", ConfigConsts.APOLLO_META_KEY, PropertiesFactory.APOLLO_PROPERTY_ORDER_ENABLE};
+  private static final Splitter NAMESPACE_SPLITTER = Splitter.on(",").omitEmptyStrings()
+      .trimResults();
+  public static final String[] APOLLO_SYSTEM_PROPERTIES = {ApolloClientSystemConsts.APP_ID,
+      ApolloClientSystemConsts.APOLLO_CLUSTER,
+      ApolloClientSystemConsts.APOLLO_CACHE_DIR,
+      ApolloClientSystemConsts.APOLLO_ACCESS_KEY_SECRET,
+      ApolloClientSystemConsts.APOLLO_META,
+      ApolloClientSystemConsts.APOLLO_CONFIG_SERVICE,
+      ApolloClientSystemConsts.APOLLO_PROPERTY_ORDER_ENABLE};
 
   private final ConfigPropertySourceFactory configPropertySourceFactory = SpringInjector
       .getInstance(ConfigPropertySourceFactory.class);