+++ apollo-client/src/main/java/com/ctrip/framework/apollo/util/factory/PropertiesFactory.java
@@ -16,6 +16,7 @@
  */
 package com.ctrip.framework.apollo.util.factory;
 
+import com.ctrip.framework.apollo.core.ApolloClientSystemConsts;
 import java.util.Properties;
 
 /**
@@ -28,7 +29,7 @@ public interface PropertiesFactory {
   /**
    * Configuration to keep properties order as same as line order in .yml/.yaml/.properties file.
    */
-  String APOLLO_PROPERTY_ORDER_ENABLE = "apollo.property.order.enable";
+  String APOLLO_PROPERTY_ORDER_ENABLE = ApolloClientSystemConsts.APOLLO_PROPERTY_ORDER_ENABLE;
 
   /**
    * <pre>