+++ apollo-client/src/test/java/com/ctrip/framework/apollo/spring/boot/ApolloApplicationContextInitializerTest.java
@@ -21,6 +21,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import com.ctrip.framework.apollo.core.ConfigConsts;
+import com.ctrip.framework.apollo.core.ApolloClientSystemConsts;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -37,9 +38,9 @@ public class ApolloApplicationContextInitializerTest {
 
   @After
   public void tearDown() throws Exception {
-    System.clearProperty("app.id");
+    System.clearProperty(ApolloClientSystemConsts.APP_ID);
     System.clearProperty(ConfigConsts.APOLLO_CLUSTER_KEY);
-    System.clearProperty("apollo.cacheDir");
+    System.clearProperty(ApolloClientSystemConsts.APOLLO_CACHE_DIR);
     System.clearProperty(ConfigConsts.APOLLO_META_KEY);
   }
 
@@ -52,16 +53,16 @@ public class ApolloApplicationContextInitializerTest {
 
     ConfigurableEnvironment environment = mock(ConfigurableEnvironment.class);
 
-    when(environment.getProperty("app.id")).thenReturn(someAppId);
+    when(environment.getProperty(ApolloClientSystemConsts.APP_ID)).thenReturn(someAppId);
     when(environment.getProperty(ConfigConsts.APOLLO_CLUSTER_KEY)).thenReturn(someCluster);
-    when(environment.getProperty("apollo.cacheDir")).thenReturn(someCacheDir);
+    when(environment.getProperty(ApolloClientSystemConsts.APOLLO_CACHE_DIR)).thenReturn(someCacheDir);
     when(environment.getProperty(ConfigConsts.APOLLO_META_KEY)).thenReturn(someApolloMeta);
 
     apolloApplicationContextInitializer.initializeSystemProperty(environment);
 
-    assertEquals(someAppId, System.getProperty("app.id"));
+    assertEquals(someAppId, System.getProperty(ApolloClientSystemConsts.APP_ID));
     assertEquals(someCluster, System.getProperty(ConfigConsts.APOLLO_CLUSTER_KEY));
-    assertEquals(someCacheDir, System.getProperty("apollo.cacheDir"));
+    assertEquals(someCacheDir, System.getProperty(ApolloClientSystemConsts.APOLLO_CACHE_DIR));
     assertEquals(someApolloMeta, System.getProperty(ConfigConsts.APOLLO_META_KEY));
   }
 
@@ -72,9 +73,9 @@ public class ApolloApplicationContextInitializerTest {
     String someCacheDir = "someCacheDir";
     String someApolloMeta = "someApolloMeta";
 
-    System.setProperty("app.id", someAppId);
+    System.setProperty(ApolloClientSystemConsts.APP_ID, someAppId);
     System.setProperty(ConfigConsts.APOLLO_CLUSTER_KEY, someCluster);
-    System.setProperty("apollo.cacheDir", someCacheDir);
+    System.setProperty(ApolloClientSystemConsts.APOLLO_CACHE_DIR, someCacheDir);
     System.setProperty(ConfigConsts.APOLLO_META_KEY, someApolloMeta);
 
     String anotherAppId = "anotherAppId";
@@ -84,16 +85,16 @@ public class ApolloApplicationContextInitializerTest {
 
     ConfigurableEnvironment environment = mock(ConfigurableEnvironment.class);
 
-    when(environment.getProperty("app.id")).thenReturn(anotherAppId);
+    when(environment.getProperty(ApolloClientSystemConsts.APP_ID)).thenReturn(anotherAppId);
     when(environment.getProperty(ConfigConsts.APOLLO_CLUSTER_KEY)).thenReturn(anotherCluster);
-    when(environment.getProperty("apollo.cacheDir")).thenReturn(anotherCacheDir);
+    when(environment.getProperty(ApolloClientSystemConsts.APOLLO_CACHE_DIR)).thenReturn(anotherCacheDir);
     when(environment.getProperty(ConfigConsts.APOLLO_META_KEY)).thenReturn(anotherApolloMeta);
 
     apolloApplicationContextInitializer.initializeSystemProperty(environment);
 
-    assertEquals(someAppId, System.getProperty("app.id"));
+    assertEquals(someAppId, System.getProperty(ApolloClientSystemConsts.APP_ID));
     assertEquals(someCluster, System.getProperty(ConfigConsts.APOLLO_CLUSTER_KEY));
-    assertEquals(someCacheDir, System.getProperty("apollo.cacheDir"));
+    assertEquals(someCacheDir, System.getProperty(ApolloClientSystemConsts.APOLLO_CACHE_DIR));
     assertEquals(someApolloMeta, System.getProperty(ConfigConsts.APOLLO_META_KEY));
   }
 
@@ -103,9 +104,9 @@ public class ApolloApplicationContextInitializerTest {
 
     apolloApplicationContextInitializer.initializeSystemProperty(environment);
 
-    assertNull(System.getProperty("app.id"));
+    assertNull(System.getProperty(ApolloClientSystemConsts.APP_ID));
     assertNull(System.getProperty(ConfigConsts.APOLLO_CLUSTER_KEY));
-    assertNull(System.getProperty("apollo.cacheDir"));
+    assertNull(System.getProperty(ApolloClientSystemConsts.APOLLO_CACHE_DIR));
     assertNull(System.getProperty(ConfigConsts.APOLLO_META_KEY));
   }
 }