+++ apollo-client/src/test/java/com/ctrip/framework/apollo/util/ConfigUtilTest.java
@@ -18,6 +18,7 @@ package com.ctrip.framework.apollo.util;
 
 import com.ctrip.framework.apollo.core.ConfigConsts;
 
+import com.ctrip.framework.apollo.core.ApolloClientSystemConsts;
 import com.ctrip.framework.apollo.util.factory.PropertiesFactory;
 import java.io.File;
 import org.junit.After;
@@ -43,7 +44,7 @@ public class ConfigUtilTest {
     System.clearProperty("apollo.configCacheSize");
     System.clearProperty("apollo.longPollingInitialDelayInMills");
     System.clearProperty("apollo.autoUpdateInjectedSpringProperties");
-    System.clearProperty("apollo.cacheDir");
+    System.clearProperty(ApolloClientSystemConsts.APOLLO_CACHE_DIR);
     System.clearProperty(PropertiesFactory.APOLLO_PROPERTY_ORDER_ENABLE);
   }
 
@@ -215,7 +216,7 @@ public class ConfigUtilTest {
     String someCacheDir = "someCacheDir";
     String someAppId = "someAppId";
 
-    System.setProperty("apollo.cacheDir", someCacheDir);
+    System.setProperty(ApolloClientSystemConsts.APOLLO_CACHE_DIR, someCacheDir);
 
     ConfigUtil configUtil = spy(new ConfigUtil());