+++ apollo-mockserver/src/main/java/com/ctrip/framework/apollo/mockserver/EmbeddedApollo.java
@@ -17,6 +17,7 @@
 package com.ctrip.framework.apollo.mockserver;
 
 import com.ctrip.framework.apollo.build.ApolloInjector;
+import com.ctrip.framework.apollo.core.ApolloClientSystemConsts;
 import com.ctrip.framework.apollo.core.dto.ApolloConfig;
 import com.ctrip.framework.apollo.core.dto.ApolloConfigNotification;
 import com.ctrip.framework.apollo.core.utils.ResourceUtils;
@@ -115,7 +116,7 @@ public class EmbeddedApollo extends ExternalResource {
   }
 
   private void mockConfigServiceUrl(String url) throws Exception {
-    System.setProperty("apollo.configService", url);
+    System.setProperty(ApolloClientSystemConsts.APOLLO_CONFIG_SERVICE, url);
 
     CONFIG_SERVICE_LOCATOR_CLEAR.invoke(CONFIG_SERVICE_LOCATOR);
   }